Persist...and then some!
