--- src/app/deck/containers/view-deck.component.html
@@ -1,9 +1,0 @@
-<bfr-nav-bar [decks]="decks$ | async" (addDeck)=onAddDeck($event)></bfr-nav-bar>
-<div class="h-100 d-flex">
-  <bfr-view-subreddit-page
-    class="subreddit-container mt-5 pt-4 w-100"
-    *ngFor="let subredditId of currentSubredditIds$ | async"
-    [subredditId]="subredditId">
-  </bfr-view-subreddit-page>
-</div>
-
