--- src/app/deck/models/deck.ts
@@ -1,5 +1,0 @@
-export interface Deck {
-  id: string;
-  name: string;
-  subredditIds: string[];
-}
