+++ src/app/decks/components/nav-bar/component.scss
