+++ src/app/decks/containers/view-deck/page.html
@@ -1,0 +1,12 @@
+<bfr-nav-bar [decks]="decks$ | async" (addDeck)=onAddDeck($event)></bfr-nav-bar>
+<div class="h-100 d-flex">
+  <bfr-view-subreddit-page
+    class="subreddit-container mt-5 pt-4 w-100"
+    *ngFor="let subredditId of currentSubredditIds$ | async"
+    [subredditId]="subredditId"
+    [deckId]="currentDeckId$ | async"
+    (setType)="onSetType($event)"
+    (setSort)="onSetSort($event)">
+  </bfr-view-subreddit-page>
+</div>
+
