+++ src/app/decks/containers/view-deck/page.scss
