+++ src/app/decks/containers/view-subreddit/page.html
@@ -1,0 +1,7 @@
+<bfr-subreddit-detail
+  [subreddit]="subreddit$ | async"
+  [posts]="posts$ | async"
+  [settings]="settings$ | async"
+  (setType)="onSetType($event)"
+  (setSort)="onSetSort($event)">
+</bfr-subreddit-detail>
