+++ src/app/decks/models/deck.ts
@@ -1,0 +1,11 @@
+export interface Deck {
+  id: string;
+  name: string;
+  subredditIds: string[];
+  subredditSettings: {
+    [id: string]: {
+      type: string,
+      sort: string
+    }
+  }
+}
