+++ src/app/decks/models/subreddit.ts
@@ -1,0 +1,6 @@
+export interface Subreddit {
+  id: string;
+  loading: boolean;
+  loaded: boolean;
+  postIds: string[];
+}
