+++ src/app/reducers/index.ts
@@ -49,5 +49,5 @@ export function logger(reducer: ActionReducer<State>): ActionReducer<State> {
  * that will be composed to form the root meta-reducer.
  */
 export const metaReducers: MetaReducer<State>[] = !environment.production
-  ? [logger, storeFreeze]
+  ? [/*logger, */storeFreeze]
   : [];
