--- src/app/subreddit/models/subreddit.ts
@@ -1,7 +1,0 @@
-export interface Subreddit {
-  id: string;
-  type: string;
-  loading: boolean;
-  loaded: boolean;
-  postIds: string[];
-}
